fix warnings
BUG=skia:

Review URL: https://codereview.chromium.org/153003005

git-svn-id: http://skia.googlecode.com/svn/trunk@13392 2bbb7eff-a529-9590-31e7-b0007b416f81

+++ src/core/SkBitmap.cpp
@@ -259,7 +259,7 @@ bool SkBitmap::setConfig(const SkImageInfo& info, size_t rowBytes) {
     if ((int32_t)mrb != mrb) {
         return reset_return_false(this);
     }
-    if ((ssize_t)rowBytes != (int32_t)rowBytes) {
+    if ((int64_t)rowBytes != (int32_t)rowBytes) {
         return reset_return_false(this);
     }
